% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nucleation_Regression.R
\name{nucl.m}
\alias{nucl.m}
\title{Probabilistic nucleation model parameter estimation from microfluidics data.}
\usage{
res <- nucl.m(1000, 100, m=0.2, k0=7e-6, k=2e-4, origin=c(0,1), filename_in="S1_exp_5_60", method="SA")
}
\arguments{
\item{nboot}{Number of boostrap samples, typically 500 to 2000.Default=1000}

\item{npoints_interp}{Number of points for continuous model display inside the experimental range. Default=100.}

\item{origin}{= c(0, 1) if data must pass through origin (time=0, f0=1) ; Optional argument}

\item{m}{Initialisation of nucleation parameter m}

\item{k0}{Initialisation of nucleation parameter k0 (s-1)}

\item{k}{Initialisation of nucleation parameter k (s-1)}

\item{filename_in}{= name of entry data file, .csv extension.
\cr Entry data file should like this:
\cr t;p
\cr 300;0.975
\cr 600;0.952
\cr...}

\item{method}{= "SA" or "LM". Default="SA"}
}
\value{
None
}
\description{
Probabilistic nucleation model parameter estimation from microfluidics data.
\cr The function takes as input a 2-column data file (see Arguments) that gives the time (in seconds) and the fraction f0 of empty droplets.
\cr Nucleation model parameters m, k0 and k are estimated by bootstrapping, where nboot is the number of boostrap samples.
\cr Parameter estimation is performed by simulated annealing or Levenberg-Marquardt, using the
GenSA and minpack.lm packages respectively.
\cr It is recommended that the user uses the "SA" option initially with nboot=1, and
then uses the output values of m, k0 and k as initialisation values for the "LM" estimation
with nboot = 1000 typically.
\cr The function outputs the results in filename_in_nboot.csv.
}
\references{
Bourgeois, F. Teychene, S. and Biscans, B., Applicability of probabilistic 
nucleation modelling for analysis of microfluidics data, KONA Powder and Particle Journal, 35, 2018.
}
\author{
Florent Bourgeois, florent.bourgeois@ensiacet.fr

Laboratoire de Genie Chimique, Toulouse, FRANCE (\url{http://www.lgc.cnrs.fr})
}
\keyword{modelling}
\keyword{nucleation}
